pyproject.toml
[tool.poetry]
name = "screen-understanding-system"
version = "0.1.0"
description = "AI-powered screen understanding system using OCR, OmniParser and VLMs"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.9"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
torchvision = "^0.16.0"
transformers = "^4.36.0"
pillow = "^10.1.0"
numpy = "^1.24.0"
opencv-python = "^4.8.0"
paddlepaddle = "^2.5.0"
paddleocr = "^2.7.0"
accelerate = "^0.25.0"
sentencepiece = "^0.1.99"
protobuf = "^4.25.0"
aiofiles = "^23.2.1"
pydantic = "^2.5.0"
tqdm = "^4.66.0"
matplotlib = "^3.8.0"
scipy = "^1.11.0"
huggingface-hub = "^0.20.0"
datasets = "^2.16.0"
einops = "^0.7.0"
timm = "^0.9.12"
nest-asyncio = "^1.5.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.8.0"
ipython = "^8.18.0"
jupyter = "^1.0.0"
notebook = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
asyncio_mode = "auto"
